flask_multi_model_api/requirements.txt
absl-py==2.2.2
astunparse==1.6.3
attrs==25.3.0
audioread==3.0.1
blinker==1.9.0
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
contourpy==1.3.2
cryptography==36.0.2
cycler==0.12.1
decorator==5.2.1
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
flatbuffers==25.2.10
fonttools==4.57.0
gast==0.6.0
google-pasta==0.2.0
greenlet==3.2.1
grpcio==1.71.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
jax==0.6.0
jaxlib==0.6.0
Jinja2==3.1.6
joblib==1.5.0
keras==3.9.2
kiwisolver==1.4.8
lazy_loader==0.4
libclang==18.1.1
librosa==0.11.0
llvmlite==0.44.0
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.9
ml_dtypes==0.5.1
msgpack==1.1.0
namex==0.0.9
numba==0.61.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.15.0
packaging==25.0
pandas==2.2.3
pillow==11.2.1
platformdirs==4.3.7
pooch==1.8.2
protobuf==4.25.7
pycparser==2.22
Pygments==2.19.1
PyMySQL==1.1.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.3
rich==14.0.0
scikit-learn==1.6.1
scipy==1.15.2
sentencepiece==0.2.0
six==1.17.0
sounddevice==0.5.1
soundfile==0.13.1
soxr==0.5.0.post1
SQLAlchemy==2.0.40
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-io-gcs-filesystem==0.37.1
termcolor==3.1.0
threadpoolctl==3.6.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
Werkzeug==3.1.3
wrapt==1.17.2
